+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.3] - 2025-07-08
+
 ### Added
 
 - Add warning to `sereto findings add` TUI when the finding already exists, and an option to overwrite it.
@@ -458,7 +460,8 @@ We registered a dummy package to PyPI to test the publishing. Therefore a versio
 Initial version
 
 
-[unreleased]: https://github.com/s3r3t0/sereto/compare/v0.3.2...HEAD
+[unreleased]: https://github.com/s3r3t0/sereto/compare/v0.3.3...HEAD
+[0.3.3]: https://github.com/s3r3t0/sereto/compare/v0.3.2...v0.3.3
 [0.3.2]: https://github.com/s3r3t0/sereto/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/s3r3t0/sereto/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/s3r3t0/sereto/compare/v0.2.9...v0.3.0